% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big-wrapper.R
\name{big_wrapper}
\alias{big_wrapper}
\title{Just a big, messy wrapper to wrap up some things so Tim can use them}
\usage{
big_wrapper(GS_input, Segs, M_meta)
}
\arguments{
\item{GS_input}{the rearranged (but not yet scrambled) genos}

\item{Segs}{the simulated segments}

\item{M_meta}{the marker meta data}
}
\description{
Just a big, messy wrapper to wrap up some things so Tim can use them
}
