% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_chrom_lengths.R
\name{check_chrom_lengths}
\alias{check_chrom_lengths}
\title{check that the chromosome lengths are acceptable given recombination rates}
\usage{
check_chrom_lengths(meta, rec)
}
\description{
Needs more documenting.
}
