% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rec_rates_formatting.R
\name{check_rec_rates_formatting}
\alias{check_rec_rates_formatting}
\title{check to make sure that the recombination rates tibble is formatted properly}
\usage{
check_rec_rates_formatting(rr)
}
\arguments{
\item{rr}{The tibble of recombination-rate bins like RecRates}
}
\description{
This is used internally when the rr object is about to be used.
}
\keyword{internal}
