% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sampleID.R
\name{check_sampleID}
\alias{check_sampleID}
\title{Check that individual IDs are in the same order in the genotype and the individual metadata tibbles.}
\usage{
check_sampleID(geno_data, ind_meta)
}
\arguments{
\item{geno_data}{a tibble with genotypes for each individual.
The tibble must consist of 2X the number of loci (each allele is in it's own column and two adjacent columns make a locus).
Rownames should be sample IDs.}

\item{ind_meta}{a tibble with meta data for individuals. It must consist of the columns (group, indiv).}
}
\value{
This function will return a message stateing whether the sample IDs for the genotype and individual metadata tibbles are in the same order
and will return an error message if each tibble has a different number of individuals.
}
\description{
This function will check that the order of individuals is the same in the genotype tibble and the individual metadata tibble.
There are two inputs. The first input is a tibble (see 'Geno' for an example) that contains the genotype information
for each individual in a row and has the individual IDs as the row names. The second input is a tibble with individual
metadata information (see 'I_meta' for an example) with the group or population name in the first column and the individual ID in the
second column. The function will also make sure that the number of individuals in each input tibble is the same.
}
\examples{

 # check that the order of sample IDs is the same in the genotype input and the individual metadata input tibble

check_sampleID(Geno,I_meta)

}
