% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeQs_from_segments.R
\name{computeQs_from_segments}
\alias{computeQs_from_segments}
\title{return the admixture fractions of sampled individuals}
\usage{
computeQs_from_segments(S, check_total_length = TRUE)
}
\arguments{
\item{S}{the tibble output from \code{\link{segregate}}}

\item{check_total_length}{TRUE means it checks the total
genome length in each individual to make sure it checks out.}
}
\description{
This operates on the output of segregate to add up the lengths
of all the segments segregated to different individuals to
thus compute the admixture fractions of each sampled
individual.
}
