% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_GSP.R
\name{create_GSP}
\alias{create_GSP}
\title{Create a GSP from user inputs about what type of hybrids from which populations to create hybrids,}
\usage{
create_GSP(pop1, pop2, F1 = F, F2 = F, F1B = F, F1B2 = F)
}
\arguments{
\item{pop1}{character name for population 1}

\item{pop2}{character name for population 2}

\item{F1}{logical indicating whether you would like to have sampled F1 hybrids in the output.}

\item{F2}{logical indicating whether you would like to have sampled F2 hybrids in the output.
Needs more documenting.}
}
\description{
what types of individuals they would like to sample
in terms of F1, F2, F1 backcross (F1B), or F1 backross 2 (F1B2) individuals.
Pop IDs must be of class character and at least one parameter from the list (F1,F2,F1B, or F1B2) must be TRUE.
Default setting for these parameters is FALSE.
}
\examples{
# create a GSP that produces F1s and F1B's from pops A and B
gsp <- create_GSP("A", "B", F1 = TRUE, F1B = TRUE)

}
