% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop-segs-down-gsp.R
\name{drop_segs_down_gsp}
\alias{drop_segs_down_gsp}
\title{High level function for dropping segments down a GSP}
\usage{
drop_segs_down_gsp(GSP, RR, Reps)
}
\arguments{
\item{GSP}{the pedigree to use for the simulation, in the format of
the package data \code{\link{GSP}}.}

\item{RR}{the recombination rates in the format of the package data
\code{\link{RecRates}}}

\item{Reps}{the number of times to do the simulation.  Different replicates
are denoted by the rep column in the output tibble.}
}
\description{
This one asks for the number of reps to do, and it also
automatically does it over chromosomes and returns the results
in a nice tidy tibble.
}
\examples{
simSegs <- drop_segs_down_gsp(GSP, RecRates, 4)


}
