% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_segments}
\alias{example_segments}
\title{Example of a segments tibble}
\format{
A tibble like that produced by the \code{segregate()} function.
}
\source{
Made from package functions
}
\description{
This is the Segments object from the gscramble tutorial.
It is included as a data object to use in the example for
the function \code{plot_simulated_chromosome_segments()}.
}
