% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_analysis.R
\name{full_analysis}
\alias{full_analysis}
\title{A full analysis of the included feral pig data}
\usage{
full_analysis()
}
\description{
This is put in here as a function so that it makes it easier to
make a dependency graph of all the different functions in
the package. It doesn't actually do anything except
call the high-level functions that one would call in
a typical full analysis. Then we can make a function dependency
graph like this:

\if{html}{\out{<div class="sourceCode">}}\preformatted{library(DependenciesGraphs)
deps <- funDependencies("package:gscramble", "full_analysis")
plot(deps)
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
library(gscramble)
library(DependenciesGraphs)
deps <- funDependencies("package:gscramble", "full_analysis")
plot(deps)
}
}
