% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Geno}
\alias{Geno}
\title{Genotype matrix of 329 individuals and 25165 SNP markers}
\format{A character matrix.}
\source{
Must add...
}
\description{
This is a data set from feral pigs.  It is a standard "two-column"
format genotype data matrix for a diploid species.
}
\details{
Need to document it a little more....
}
