% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gsp3}
\alias{gsp3}
\title{Tibble holding specification for a 5 member genomic permutation pedigree.}
\format{A tibble}
\source{
Tim made it up!
}
\description{
This has 3 founders
}
\details{
Need to document it a little more....
}
