% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gsp4}
\alias{gsp4}
\title{Tibble holding specification for a 7 member genomic permutation pedigree.}
\format{
A tibble
}
\source{
Created by the developers.
}
\description{
This has 4 founders, each one from a different population, and it provides
four samples that are the product of F1 matings (an A x B F1 crossing with
a C x D F1.)
}
\details{
For details on the columns, see the documentation for \code{\link{GSP}}.
}
