% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSP_opts}
\alias{GSP_opts}
\title{Tibble holding specification for a 5 member genomic permutation pedigree.}
\format{
A list of 15 tibbles
}
\source{
Rachael and Tim
}
\description{
This has 3 founders
}
\details{
Need to document it a little more....
}
