% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{I_meta}
\alias{I_meta}
\title{Meta data for 329 individuals}
\format{
A tibble with three columns: chrom, pos, and variant_id.
}
\source{
Must add...
}
\description{
Individuals IDs and group specifications for 329 individuals
whose genotypes are in \code{\link{Geno}}
}
