% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{M_meta}
\alias{M_meta}
\title{Meta data for 25165 molecular markers}
\format{
A tibble with two columns: group and indiv.
}
\source{
Must add...
}
\description{
Chromosome, position, and ID for 25156 molecular
markers in \code{\link{Geno}}
}
