% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_subscript_matrix.R
\name{make_subscript_matrix}
\alias{make_subscript_matrix}
\title{makes a two-column matrix for subscripting alleles out of a genotype matrix}
\usage{
make_subscript_matrix(n, chrom, start, end, abs_column, m_list, num_markers)
}
\description{
I need to fully document this still.
}
