% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_gs_by_pops.R
\name{perm_gs_by_pops}
\alias{perm_gs_by_pops}
\title{Take the output of rearrange_genos and permute everyone by population}
\usage{
perm_gs_by_pops(GS, preserve_haplotypes = FALSE)
}
\arguments{
\item{GS}{the tibble that is the output from rearrange_genos}

\item{preserve_haplotypes}{If true then the Geno data is assumed phased
(first allele at an individual on one haplotype and second allele on the
other) and those haplotypes are preserved in this permutation of
genomic material amongst the founders.}
}
\description{
This was put together pretty hastily
}
