% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simulated_chromosome_segments.R
\name{plot_simulated_chromomsome_segments}
\alias{plot_simulated_chromomsome_segments}
\title{Plot the simulated chromosomes of an individual}
\usage{
plot_simulated_chromomsome_segments(
  Segs,
  RR = NULL,
  rel_heights = c(chrom_ht = 4, chrom_gap = 0.8, spark_gap = 0.6 * !is.null(RR),
    spark_box = 2.2 * !is.null(RR), unit_gap = 4),
  bottom_gap = 0.3
)
}
\arguments{
\item{Segs}{a tibble of segments}

\item{RR}{a tibble of recombination rates in bins in the format
of \link{RecRates}. If this is
included, the recombination rates in cM/Mb are plotted atop the chromosomes
as a little sparkline. If it is not included, then the there are
no little sparklines above the chromsomes.}

\item{rel_heights}{a vector the the relative heights of the different
elements of each chromosomal unit of the plot.  This is a named vector
with the following elements, listed in order of the bottom of each
unit to the top:
\itemize{
\item \code{chrom_ht}: the height of the bars for each of the two chromosomes of the
pair in a chromosome unit.
\item \code{chrom_gap}: The gap between the two homologous chromosomes of the individual.
\item \code{spark_gap}: the gap between the top chromosome and the sparkline box
for recombination rates.
\item \code{spark_box}: height of the box within which the sparkline goes.  Note that
the sparkline itself will be scaled so that the highest rate anywhere withing
the genome will correspond to the top of the spark box.
\item \code{unit_gap}: The relative height of the gap between one chromosome unit and
the next.
}}

\item{bottom_gap}{the y value of the bottom chromosome unit.  Basically the
absolute distance between the y=0 line and the start of the plotted material.
Should typically be between 0 and 1.}
}
\description{
This function uses the information in the tibble about segments
dropped down a genome simulation pedigree to plot the
chromomomes of an individual colored by either
the population of origin of each segment or the founding
haplotype of origin of each segment.
}
\examples{
# must add some
}
