% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_gsp_for_hap_dropping.R
\name{prep_gsp_for_hap_dropping}
\alias{prep_gsp_for_hap_dropping}
\title{Take a gsp in tibble form and make a list suitable for gene dropping}
\usage{
prep_gsp_for_hap_dropping(gsp)
}
\arguments{
\item{gsp}{A tibble that holds the gsp}
}
\description{
Just a simple function that makes a list-based data structure
that makes it easy to gene drop chromosome segments down the
gsp.  The basic idea is to get everyone into a data structure
in which they are ordered such that by the time you get to segregating
genes from anyone, you already have the genes segregated into them.
This works by each individual having a list of gametes (post-recombination)
coming out of them, and a list of "uniters" which are the gametes coming
into them.  We just point the uniters to gametes in the previous generation
and then make sure that we shuffle the order of the gametes when they come
out of someone.
}
