% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RecRates}
\alias{RecRates}
\title{Recombination rate data for many roughly 1 Mb bins}
\format{
A tibble with four columns: chrom, chrom_len, start_pos, end_pos, and rec_prob.
}
\source{
Must add, inluding how I computed it from Torterau.
}
\description{
Chromosome, start position and end position and probability of recombination
within the bin for chromosomes in pigs.
}
