% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renumber_GSP.R
\name{renumber_GSP}
\alias{renumber_GSP}
\title{Renumber GSP members by adding a constant to each}
\usage{
renumber_GSP(G, add)
}
\arguments{
\item{G}{a GSP tibble}

\item{add}{amount to add to each label}
}
\value{
Returns a GSP just like the input, but with the identity numbers
of the individuals in it incremented by \code{add}.
}
\description{
This function assumes that all individuals are named as
numerics and that their haplotypes in hap1 and hap2
are named Xa and Xb,respectively, and their
samples are named sX, where X is an integer,
}
\examples{
# get an example GSP
G <- create_GSP(pop1 = "p1", pop2 = "p2", F1B2 = TRUE)


}
