% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_level_founder_haplos.R
\name{sim_level_founder_haplos}
\alias{sim_level_founder_haplos}
\title{computes the simulation-level founder haplotype index for each founder haplo}
\usage{
sim_level_founder_haplos(S)
}
\arguments{
\item{S}{tibble of segments like that produced by \code{\link{segregate}}.}
}
\description{
This takes the output of segregate and deals with the multiple gpp's and reps
to come up with a unique index for each found haplotype, so that those haplotypes
can all, eventually, be accessed easily out of the genotype matrix.
}
\details{
Note that the result one gets here is basically a the output of segregate with
an additional column added to it:  sim_level_founder_haplo.  This is the index
of the haplotype within each group_origin that should be used.

Along the way, this function does some light checking to make sure that the
rs_founder_haplo values are dense with gpp and rep as they should be.
}
