% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_up_sampled_haplos.R
\name{tidy_up_sampled_haplos}
\alias{tidy_up_sampled_haplos}
\title{takes the waka_waka table and returns a simple tibble with results for each sampled "hybridized" gamete from each chromosome
#'
Not yet documented}
\usage{
tidy_up_sampled_haplos(W)
}
\arguments{
\item{W}{a tibble.  see how it is used in code below.}
}
\description{
takes the waka_waka table and returns a simple tibble with results for each sampled "hybridized" gamete from each chromosome
#'
Not yet documented
}
\keyword{internal}
