% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xover.R
\name{xover}
\alias{xover}
\title{internal function to do crossovers and create two recombinations}
\usage{
xover(V1, V2, R)
}
\arguments{
\item{V1}{integer vector of breakpoints already existing on the
first incoming gamete.  Its names are the names of the
founder haplotype that the left end originates from (i.e. it is from the named
haplotype up until it changes at each point).  For example
c(A = 0, B = 12890, B = 30000) would work for a 30 Kb chromosome in which there
is a single recombination at 12890.  Note that these vectors have to have a first
value of 0 and a final value of the chromosome length.}

\item{V2}{integer vector of breakpoints of the second incoming gamete.  Format
is just like it is for V1.}

\item{R}{a vector of new breakpoints to insert into the existing ones on each gamete.}
}
\value{
This sends back two updated gametes, V1 and V2, but with the new points
of recombination stuck in there.  Note, for two incoming gametes there are two
outgoing gametes, but we aren't "re-using" any genomic sequence.
}
\description{
This doesn't choose the recombination points.  That has to be done
and the results passed into this function.
}
