% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_GSP.R
\name{create_GSP}
\alias{create_GSP}
\title{Creates a GSP based on user inputs for which populations from which to create hybrids, what types of individuals they would like to sample
in terms of F1, F2, F1 backcross (F1B), or F1 backross 2 (F1B2) individuals.
Pop IDs must be of class character and at least one parameter from the list (F1,F2,F1B, or F1B2) must be TRUE.
Default setting for these parameters is FALSE.}
\usage{
create_GSP(pop1, pop2, F1 = F, F2 = F, F1B = F, F1B2 = F)
}
\description{
Needs more documenting.
}
