% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSP}
\alias{GSP}
\title{Tibble holding specification for a 13 member genomic permutation pedigree}
\format{
A tibble
}
\source{
I just made it up!
}
\description{
This is just an example.
}
\details{
Need to document it a little more....
}
