% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gsp3}
\alias{gsp3}
\title{Tibble holding specification for a 5 member genomic permutation pedigree.}
\format{
A tibble
}
\source{
Created by the developers.
}
\description{
This has 3 founders
}
\details{
For details on the columns, see the documentation for \code{\link{GSP}}.
}
