% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSP_opts}
\alias{GSP_opts}
\title{A list of tibbles specifying the pedigrees available from \code{createGSP()}}
\format{
A list of 15 tibbles
}
\source{
Written by package authors Rachael and Tim
}
\description{
This is a list that is used by the function \code{createGSP()}.  There are 15
different genomic simulation pedigrees, specified as tibbles, in this
list.
}
