% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{I_meta}
\alias{I_meta}
\title{Metadata for 78 individuals}
\format{
A tibble with two columns: group and indiv.
}
\source{
USDA-APHIS-WS-National Wildlife Research Center Feral Swine Genetic Archive
}
\description{
The tibble 'I_meta' contains the Individuals IDs and group specifications for 78 individuals. Each row of 'I_meta'
contains the metadata for an individual. 'I_meta' has two columns, a 'group' column with the group or population assignment
for a given individual and an 'indiv' column with the individual sample IDs. The information in each column are stored as characters.
The order of the individual IDs corresponds to the genotypes found in \code{\link{Geno}}.
}
