% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_gs_by_pops.R
\name{mat_scramble}
\alias{mat_scramble}
\title{Assumes that M is a matrix with L rows (number of markers) and
2 * N (N = number of individuals) columns.  Holds position of
missing data constant.}
\usage{
mat_scramble(M, missing = "0")
}
\description{
Assumes that M is a matrix with L rows (number of markers) and
2 * N (N = number of individuals) columns.  Holds position of
missing data constant.
}
