% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_gs_by_pops.R
\name{perm_gs_by_pops}
\alias{perm_gs_by_pops}
\title{Take the output of rearrange_genos and permute everyone by population}
\usage{
perm_gs_by_pops(GS)
}
\arguments{
\item{GS}{the tibble that is the output from rearrange_genos}
}
\description{
This was put together pretty hastily
}
