% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recom_point.R
\name{recomb_point}
\alias{recomb_point}
\title{chooses positions of recombination points}
\usage{
recomb_point(M, always_one = TRUE)
}
\arguments{
\item{M}{a tibble that has the columns start_pos, end_pos, and theta (where theta is the
probability (or intensity) of recombinations in the interval defined by start_pos and end_pos.}
}
\description{
It is based on a map M.
}
\keyword{internal}
