% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg2tib.R
\name{seg2tib}
\alias{seg2tib}
\title{Takes a gamete in segment format and returns a tibble with Pop and indiv_index}
\usage{
seg2tib(s)
}
\arguments{
\item{s}{a gamete in segment format}
}
\value{
Returns a tibble with columns \code{tmp_seg_names}, \code{start}, and \code{end},
that show the origin (in \code{tmp_seg_names}) of segments that start at \code{start}
and end at \code{end}.
}
\description{
A small helper function.
}
\examples{
# first make a segment that has pieces from a few different founders
 V <- c(Amy = 0, Bob = 10000, Joe = 30000, Frieda = 40000)
 seg2tib(V)
}
