% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments2markers.R
\name{segments2markers}
\alias{segments2markers}
\title{Just a big, messy wrapper to wrap up some things so Tim can use them}
\usage{
segments2markers(Segs, Im, Mm, G)
}
\arguments{
\item{Segs}{the simulated segments}

\item{Im}{the individual meta data, like that in \code{\link{I_meta}}}

\item{Mm}{the marker meta data formatted like that in \code{\link{M_meta}}}

\item{G}{the marker genotype data as a matrix like \code{\link{Geno}}}
}
\description{
Just a big, messy wrapper to wrap up some things so Tim can use them
}
