% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segregate.R
\name{segregate}
\alias{segregate}
\title{must document this one fully}
\usage{
segregate(request, RR)
}
\arguments{
\item{request}{a tibble with list columns "gpp" and "reppop".}

\item{RR}{the recombination rates in the format of the package data
\code{\link{RecRates}}}
}
\description{
Gonna do that later
}
